Move component builders to packages
