+++ src/main/java/com/freya02/botcommands/api/components/Components.java
@@ -4,7 +4,13 @@ import com.freya02.botcommands.api.BContext;
 import com.freya02.botcommands.api.Logging;
 import com.freya02.botcommands.api.components.annotations.JDAButtonListener;
 import com.freya02.botcommands.api.components.annotations.JDASelectionMenuListener;
-import com.freya02.botcommands.api.components.builder.*;
+import com.freya02.botcommands.api.components.builder.ComponentBuilder;
+import com.freya02.botcommands.api.components.builder.buttons.LambdaButtonBuilder;
+import com.freya02.botcommands.api.components.builder.buttons.PersistentButtonBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaEntitySelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaStringSelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.PersistentEntitySelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.PersistentStringSelectionMenuBuilder;
 import com.freya02.botcommands.api.components.event.ButtonEvent;
 import com.freya02.botcommands.api.components.event.StringSelectionEvent;
 import com.freya02.botcommands.internal.utils.Utils;
