+++ src/main/java/com/freya02/botcommands/api/components/DefaultComponentManager.java
@@ -1,7 +1,14 @@
 package com.freya02.botcommands.api.components;
 
 import com.freya02.botcommands.api.Logging;
-import com.freya02.botcommands.api.components.builder.*;
+import com.freya02.botcommands.api.components.builder.ComponentBuilder;
+import com.freya02.botcommands.api.components.builder.LambdaComponentTimeoutInfo;
+import com.freya02.botcommands.api.components.builder.PersistentComponentBuilder;
+import com.freya02.botcommands.api.components.builder.PersistentComponentTimeoutInfo;
+import com.freya02.botcommands.api.components.builder.buttons.LambdaButtonBuilder;
+import com.freya02.botcommands.api.components.builder.buttons.PersistentButtonBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaSelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.PersistentSelectionMenuBuilder;
 import com.freya02.botcommands.internal.components.HandleComponentResult;
 import com.freya02.botcommands.internal.components.data.LambdaButtonData;
 import com.freya02.botcommands.internal.components.data.LambdaSelectionMenuData;
