--- src/main/java/com/freya02/botcommands/api/components/builder/PersistentSelectionMenuBuilder.java
@@ -1,7 +1,0 @@
-package com.freya02.botcommands.api.components.builder;
-
-public interface PersistentSelectionMenuBuilder<T extends PersistentSelectionMenuBuilder<T>>
-		extends ComponentBuilder<T>,
-		        PersistentComponentBuilder<T> {
-
-}
