+++ src/main/java/com/freya02/botcommands/api/components/builder/selects/LambdaStringSelectionMenuBuilder.java
@@ -1,27 +1,28 @@
-package com.freya02.botcommands.api.components.builder;
+package com.freya02.botcommands.api.components.builder.selects;
 
 import com.freya02.botcommands.api.BContext;
 import com.freya02.botcommands.api.components.ComponentManager;
-import com.freya02.botcommands.api.components.EntitySelectionConsumer;
 import com.freya02.botcommands.api.components.InteractionConstraints;
-import com.freya02.botcommands.api.components.event.EntitySelectionEvent;
+import com.freya02.botcommands.api.components.StringSelectionConsumer;
+import com.freya02.botcommands.api.components.builder.LambdaComponentTimeoutInfo;
+import com.freya02.botcommands.api.components.event.StringSelectionEvent;
 import com.freya02.botcommands.internal.utils.Utils;
-import net.dv8tion.jda.api.interactions.components.selections.EntitySelectMenu;
+import net.dv8tion.jda.api.interactions.components.selections.StringSelectMenu;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.TimeUnit;
 
-public class LambdaEntitySelectionMenuBuilder
-		extends EntitySelectMenu.Builder
-		implements LambdaSelectionMenuBuilder<LambdaEntitySelectionMenuBuilder, EntitySelectionEvent> {
+public class LambdaStringSelectionMenuBuilder
+		extends StringSelectMenu.Builder
+		implements LambdaSelectionMenuBuilder<LambdaStringSelectionMenuBuilder, StringSelectionEvent> {
 	private final BContext context;
-	private final EntitySelectionConsumer consumer;
+	private final StringSelectionConsumer consumer;
 
 	private boolean oneUse;
 	private LambdaComponentTimeoutInfo timeoutInfo = new LambdaComponentTimeoutInfo(0, TimeUnit.MILLISECONDS, () -> {});
 	private final InteractionConstraints interactionConstraints = new InteractionConstraints();
 
-	public LambdaEntitySelectionMenuBuilder(BContext context, EntitySelectionConsumer consumer) {
+	public LambdaStringSelectionMenuBuilder(BContext context, StringSelectionConsumer consumer) {
 		super("fake");
 
 		this.context = context;
@@ -30,13 +31,13 @@ public class LambdaEntitySelectionMenuBuilder
 
 	@NotNull
 	@Override
-	public EntitySelectionConsumer getConsumer() {
+	public StringSelectionConsumer getConsumer() {
 		return consumer;
 	}
 
 	@NotNull
 	@Override
-	public EntitySelectMenu build() {
+	public StringSelectMenu build() {
 		final ComponentManager componentManager = Utils.getComponentManager(context);
 
 		setId(componentManager.putLambdaSelectMenu(this));
@@ -45,14 +46,14 @@ public class LambdaEntitySelectionMenuBuilder
 	}
 
 	@Override
-	public LambdaEntitySelectionMenuBuilder oneUse() {
+	public LambdaStringSelectionMenuBuilder oneUse() {
 		this.oneUse = true;
 
 		return this;
 	}
 
 	@Override
-	public LambdaEntitySelectionMenuBuilder timeout(long timeout, @NotNull TimeUnit timeoutUnit, @NotNull Runnable timeoutCallback) {
+	public LambdaStringSelectionMenuBuilder timeout(long timeout, @NotNull TimeUnit timeoutUnit, @NotNull Runnable timeoutCallback) {
 		this.timeoutInfo = new LambdaComponentTimeoutInfo(timeout, timeoutUnit, timeoutCallback);
 
 		return this;
