+++ src/main/java/com/freya02/botcommands/api/components/builder/selects/PersistentSelectionMenuBuilder.java
@@ -1,0 +1,10 @@
+package com.freya02.botcommands.api.components.builder.selects;
+
+import com.freya02.botcommands.api.components.builder.ComponentBuilder;
+import com.freya02.botcommands.api.components.builder.PersistentComponentBuilder;
+
+public interface PersistentSelectionMenuBuilder<T extends PersistentSelectionMenuBuilder<T>>
+		extends ComponentBuilder<T>,
+		        PersistentComponentBuilder<T> {
+
+}
