+++ src/test/java/com/freya02/botcommands/test/commands/slash/SlashSelectMenu.java
@@ -6,9 +6,9 @@ import com.freya02.botcommands.api.application.annotations.AppOption;
 import com.freya02.botcommands.api.application.slash.GuildSlashEvent;
 import com.freya02.botcommands.api.application.slash.annotations.JDASlashCommand;
 import com.freya02.botcommands.api.components.Components;
-import com.freya02.botcommands.api.components.builder.LambdaEntitySelectionMenuBuilder;
-import com.freya02.botcommands.api.components.builder.LambdaSelectionMenuBuilder;
-import com.freya02.botcommands.api.components.builder.LambdaStringSelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaEntitySelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaSelectionMenuBuilder;
+import com.freya02.botcommands.api.components.builder.selects.LambdaStringSelectionMenuBuilder;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.channel.ChannelType;
